Refactor path macro extensions
