+++ Cargo.toml
@@ -11,7 +11,7 @@ members = [
 [features]
 swagger_ui = ["rust-embed", "mime_guess"]
 debug = ["utoipa-gen/debug"]
-actix_gen = ["utoipa-gen/actix_gen"]
+actix_extras = ["utoipa-gen/actix_extras"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
