+++ tests/utoipa_gen_test.rs
@@ -26,7 +26,7 @@ struct Foo {
         ("ids" = [i32], query, description = "Search foos by ids"),
    ]
 )]
-#[get("/foo")]
+#[get("/foo/{name:.*}")]
 // #[deprecated = "this is deprecated"]
 // web::Path(id): web::Path<i32>
 async fn foo_delete(web::Query(foo): web::Query<Foo>) -> impl Responder {
