+++ utoipa-gen/Cargo.toml
@@ -16,4 +16,4 @@ proc-macro-error = "1.0.4"
 
 [features]
 debug = []
-actix_gen = []
+actix_extras = []
